+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added helper methods for keeping track of cell/range dependencies: getCellPrecedents and getCellDependents. (#441)
 - Added 4 financial functions FV, PMT, PPMT, IPMT.
 - Added FORMULATEXT function.
-- Added 2 date functions: WEEKDAY, DATEVALUE.
+- Added 5 date functions: WEEKDAY, DATEVALUE, HOUR, MINUTE, SECOND.
 
 ### Changed
 - Operation `moveCells` creating cyclic dependencies does not cause losing original formula. (#479)